+++ activerecord/lib/active_record.rb
@@ -53,17 +53,6 @@ module ActiveRecord
   autoload :ReadonlyAttributes
   autoload :Reflection
   autoload :Sanitization
-
-  # ActiveRecord::SessionStore depends on the abstract store in Action Pack.
-  # Eager loading this class would break client code that eager loads Active
-  # Record standalone.
-  #
-  # Note that the Rails application generator creates an initializer specific
-  # for setting the session store. Thus, albeit in theory this autoload would
-  # not be thread-safe, in practice it is because if the application uses this
-  # session store its autoload happens at boot time.
-  autoload :SessionStore
-
   autoload :Schema
   autoload :SchemaDumper
   autoload :SchemaMigration